requirements.txt
Django==3.1.5
django-mysql==3.11.1
djangorestframework==3.12.2
mysqlclient==2.0.3
PyMySql==1.0.2
PyJWT==2.0.1
django-cors-headers==3.5.0
